Also recover type for errors in static items in type_of
The const interner uses the body to intern static allocations.
This means that an allocation will be mutable if the resulting type
contains interior mutability.
Const eval contains an assertion that types that are `Freeze` do not
have mutable allocations, which was failing for cases where the type was
a type error.

An alternative fix for this would be to avoid the assertion in const
eval for type errors, allowing those to both be mutable and immutable.
This is implemented in a follow-up commit as well.