Better recover static/const type in type_of
When the type is an inferred placeholder or an error, we used to infer
the type for an error message and then return `TyKind::Error`.
Now we return the proper type, which helps follow-up code do more
checking.

The primary motivation for this was fixing the const-eval mutability
assertion bug, because this commit forwards the inferred type
to that assertion, but it is also just nicer in general.