Remove the `missing_copy_implementations` lint.
This lint made a certain amount of sense for people writing code during
the period leading up to to Rust 1.0. It doesn't make much sense today.
There are many valid cases of types that could impl `Copy` but do not.
The lint is extremely low value and moderately complicated. This commit
removes it.