unify git command preperation
Due to https://github.com/rust-lang/rust/issues/125954, we had to modify git invocations with
certain flags in https://github.com/rust-lang/rust/pull/126255. However, because there are so many
instances of `Command::new("git")` in bootstrap, it is difficult to apply these solutions to all of
them.

This PR creates a helper function that unifies the git usage in bootstrap. Meaning, whenever special flags
or hacks are needed, we can apply them to this single function which makes things much simpler for the bootstrap team.

Signed-off-by: onur-ozkan <work@onurozkan.dev>